Menu Bars Added

+++ src/main/java/dataTier/ReimbursementDAO.java
@@ -37,7 +37,7 @@ public class ReimbursementDAO {
 
 		while (rs.next()) {
 
-			Reimbursement reImburse = new Reimbursement(rs.getInt(1), rs.getInt(2), rs.getTimestamp(3),
+			Reimbursement reImburse = new Reimbursement(rs.getInt(1), rs.getDouble(2), rs.getTimestamp(3),
 					rs.getTimestamp(4), rs.getString(5), rs.getString(6), rs.getString(7), rs.getString(8), rs.getString(9));
 			reimbList.add(reImburse);
 			System.out.println(reImburse);
@@ -65,7 +65,7 @@ public class ReimbursementDAO {
 		List<Reimbursement> reimbList = new ArrayList<Reimbursement>();
 
 		while (rs.next()) {
-			Reimbursement reImburse = new Reimbursement(rs.getInt(1), rs.getInt(2), rs.getTimestamp(3),
+			Reimbursement reImburse = new Reimbursement(rs.getInt(1), rs.getDouble(2), rs.getTimestamp(3),
 					rs.getTimestamp(4), rs.getString(5), rs.getString(6), rs.getString(7), rs.getString(8),
 					rs.getString(9));
 			reimbList.add(reImburse);
